requirements.txt
torch
transformers
pandas
scikit-learn
shap
beautifulsoup4
newspaper3k
matplotlib
seaborn
tqdm
ipywidgets
